langchain/autogpt-gradio/requirements.txt
simplejson
requests
PyYAML
loguru
openai
langchain
langchain_openai
langchain-community
langchain_experimental
faiss-cpu
gradio